+++ spec/ideal_mollie_spec.rb
@@ -17,49 +17,57 @@ describe IdealMollie do
         banks.count > 0
 
         bank = banks.first
-        bank[:id].should eq "0031"
-        bank[:name].should eq "ABN AMRO"
+        bank.id.should eq "0031"
+        bank.name.should eq "ABN AMRO"
       end
     end
   end
 
-  context "#new_payment" do
-    it "should return a transaction hash with the correct values" do
-      VCR.use_cassette("new_payment") do
-        payment = IdealMollie.new_payment(1000, "test", "0031")
-        payment[:transaction_id].should eq "c9f93e5c2bd6c1e7c5bee5c5580c6f83"
-        payment[:amount].should eq 1000
-        payment[:currency].should eq "EUR"
-        payment[:url].should eq "https://www.abnamro.nl/nl/ideal/identification.do?randomizedstring=8433910909&trxid=30000217841224"
-        payment[:message].should eq "Your iDEAL-payment has successfully been setup. Your customer should visit the given URL to make the payment"
+  context "#new_order" do
+    it "should return a Order with the correct values" do
+      VCR.use_cassette("new_order") do
+        order = IdealMollie.new_order(1000, "test", "0031")
+        order.transaction_id.should eq "c9f93e5c2bd6c1e7c5bee5c5580c6f83"
+        order.amount.should eq 1000
+        order.currency.should eq "EUR"
+        order.url.should eq "https://www.abnamro.nl/nl/ideal/identification.do?randomizedstring=8433910909&trxid=30000217841224"
+        order.message.should eq "Your iDEAL-payment has successfully been setup. Your customer should visit the given URL to make the payment"
       end
+    end
+    it "should override the return url when specified" do
+      params = IdealMollie.new_order_params(1200, "test", "0031")
+      params[:returnurl].should eq "http://example.org/return"
+
+      params = IdealMollie.new_order_params(1200, "test", "0031", "http://another.example.org/return")
+      params[:returnurl].should eq "http://another.example.org/return"
     end
   end
 
-  context "#check_payment" do
-    it "should return a payment response with the correct values" do
-      VCR.use_cassette("check_payment") do
-        payment = IdealMollie.check_payment("c9f93e5c2bd6c1e7c5bee5c5580c6f83")
-        payment[:transaction_id].should eq "c9f93e5c2bd6c1e7c5bee5c5580c6f83"
-        payment[:amount].should eq 1000
-        payment[:currency].should eq "EUR"
-        payment[:payed].should eq false
-        payment[:message].should eq "This iDEAL-order wasn't payed for, or was already checked by you. (We give payed=true only once, for your protection)"
-        payment[:status].should eq "CheckedBefore"
+  context "#check_order" do
+    it "should return a OrderResult with the correct values" do
+      VCR.use_cassette("check_order") do
+        order_result = IdealMollie.check_order("c9f93e5c2bd6c1e7c5bee5c5580c6f83")
+        order_result.transaction_id.should eq "c9f93e5c2bd6c1e7c5bee5c5580c6f83"
+        order_result.amount.should eq 1000
+        order_result.currency.should eq "EUR"
+        order_result.payed.should eq false
+        order_result.message.should eq "This iDEAL-order wasn't payed for, or was already checked by you. (We give payed=true only once, for your protection)"
+        order_result.status.should eq "CheckedBefore"
       end
     end
-    it "should mark the transaction as payed and contain the consumer information when called by mollie" do
-      VCR.use_cassette("check_payment") do
-        payment = IdealMollie.check_payment("482d599bbcc7795727650330ad65fe9b")
-        payment[:transaction_id].should eq "482d599bbcc7795727650330ad65fe9b"
-        payment[:amount].should eq 1000
-        payment[:currency].should eq "EUR"
-        payment[:payed].should eq true
-        payment[:message].should eq "This iDEAL-order has successfuly been payed for, and this is the first time you check it."
-        consumer = payment[:consumer]
-        consumer[:name].should eq "Hr J Janssen"
-        consumer[:account].should eq "P001234567"
-        consumer[:city].should eq "Amsterdam"
+
+    it "should mark the OrderResult as payed and contain the customer information when called by mollie" do
+      VCR.use_cassette("check_order") do
+        order_result = IdealMollie.check_order("482d599bbcc7795727650330ad65fe9b")
+        order_result.transaction_id.should eq "482d599bbcc7795727650330ad65fe9b"
+        order_result.amount.should eq 1000
+        order_result.currency.should eq "EUR"
+        order_result.payed.should eq true
+        order_result.message.should eq "This iDEAL-order has successfuly been payed for, and this is the first time you check it."
+
+        order_result.customer_name.should eq "Hr J Janssen"
+        order_result.customer_account.should eq "P001234567"
+        order_result.customer_city.should eq "Amsterdam"
       end
     end
   end