+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- **26 January 2012**: version 0.0.3
+    - Added `Bank`, `Order` and `OrderResult`. Should be easier then hashes.
+    - Added example to the README
+    - Now you can override the return_url within the controller
+
 - **25 January 2012**: version 0.0.2
     - Dependecies fixed
     - Tests fixed